% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{draw_hist_boxplots}
\alias{draw_hist_boxplots}
\title{Return a layout of (nrow, ncol) plots composed of histograms with corresponding boxplot below, for each variable.}
\usage{
draw_hist_boxplots(
  data,
  nrow = 2,
  ncol = 2,
  boxframe = T,
  color = rgb(0, 0, 0)
)
}
\arguments{
\item{data}{a dataframe containing the data}

\item{nrow}{the number of rows, default is 2}

\item{ncol}{the number of colums, default is 2}

\item{boxframe}{does the boxplot have a border, default is true}

\item{color}{the color of the bar in the histogram, default is black}
}
\description{
Return a layout of (nrow, ncol) plots composed of histograms with corresponding boxplot below, for each variable.
}
\keyword{boxplot}
\keyword{draw}
\keyword{histogram}
\keyword{lioche}
\keyword{rcpack}
