% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{remove_ggplot_background}
\alias{remove_ggplot_background}
\title{A small function which return a ggplot instance but without background}
\usage{
remove_ggplot_background(plot)
}
\arguments{
\item{plot}{a ggplot instance}
}
\description{
A small function which return a ggplot instance but without background
}
\keyword{background}
\keyword{ggplot}
\keyword{lioche}
\keyword{rcpack}
