% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_on_sample_prediction.R
\name{ts.on.sample.prediction}
\alias{ts.on.sample.prediction}
\title{On sample predictions for TS}
\usage{
ts.on.sample.prediction(
  data,
  order,
  seasonal = list(order = c(0, 0, 0), period = NA),
  n.lag = 1
)
}
\arguments{
\item{data}{the (univariate) time series. Must be class of ts().}

\item{order}{the order of the time series, for the model.}

\item{seasonal}{the seasonal order of the time series, for the model.}

\item{n.lag}{the n.ahead steps to be predicted with predict().}
}
\value{
Returns a list with
\itemize{
\item pred : the 20\% predicted points
\item se : the values of each squared errors
\item mse : the mean of se, aka mean squared errors : sum(se)/n
}
}
\description{
Calculates the mean squared error of n.lag steps
predicted for a time series modeled according to a (S)AR(I)MA.
}
\details{
For the last 20\% data of the ts, compute a model with i+1 data points
and predict n.lag steps ahead, up to the length of the ts.
}
\note{
Be careful not to set the orders value too high.
The data must be the original one, not differentiated.
The predicted values returned are the first step predicted, independent of n.lag
}
\examples{
ts.on.sample.prediction(gas, order = c(1, 1, 1), seasonal = list(order = c(1, 1, 1), period = 12))

}
\author{
Lionel Lamy
}
\concept{OneAhead redesigned}
