% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_sarima_model_selection.R
\name{ts.sarima.model.selection}
\alias{ts.sarima.model.selection}
\title{Sarima model selection}
\usage{
ts.sarima.model.selection(
  data,
  max.pq = c(1, 1),
  max.PQ = c(1, 1),
  d = 1,
  D = 1,
  top = 3,
  return.best = F
)
}
\arguments{
\item{data}{the (univariate) time series. Must be class of ts().}

\item{max.pq}{vector of length 2, maximum orders for p and q.}

\item{max.PQ}{vector of length 2, maximum orders for P and Q.}

\item{d}{number of diff of lag 1 for trend elimination.}

\item{D}{number of diff of lag 'season' for seasonality elimination.}

\item{top}{the number of best models returned.}

\item{return.best}{(bool) : is the best AIC model returned.}
}
\value{
the model with the lowest AIC.
}
\description{
Comparison using the AIC criterion of different S-ARIMA models
fitted to a time series with trend and seasonality removal.
}
\details{
Displays the top x models in the console.
Given info are AIC, AICR (relative), BIC and the nb of parameters.
Returns the model with the lowest AIC if return.best is true
}
\note{
Be careful not to set the maximum values too high.
The data must be the original one, not differentiated.
}
\author{
Lionel Lamy
}
\concept{Comp.Sarima redesigned}
